requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
openpyxl==3.1.5
sqlparse==0.5.1
